# apps/backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Game Zone Management System API",
    "description": "REST API for managing a gaming cafe: users, devices, plans, sessions, transactions, and inventory.",
    "contact": {
      "name": "Gaming Cafe Team"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/plans": {
      "get": {
        "tags": [
          "plans"
        ],
        "operationId": "list_plans",
        "responses": {
          "200": {
            "description": "List plans",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanPaginationEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "plans"
        ],
        "operationId": "create_plan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Create plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/plans/active": {
      "get": {
        "tags": [
          "plans"
        ],
        "operationId": "get_active_plans",
        "responses": {
          "200": {
            "description": "List active plans",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivePlansEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/plans/{id}": {
      "get": {
        "tags": [
          "plans"
        ],
        "operationId": "get_plan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Plan ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "plans"
        ],
        "operationId": "delete_plan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Plan ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Soft-deactivated (isActive=false)"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "plans"
        ],
        "operationId": "update_plan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Plan ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivePlansEnvelope": {
        "type": "object",
        "required": [
          "success",
          "statusCode",
          "timestamp",
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Plan"
            }
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "CreatePlanDto": {
        "type": "object",
        "required": [
          "name",
          "price",
          "planType"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "deviceSubType": {
            "type": [
              "string",
              "null"
            ]
          },
          "deviceType": {
            "type": [
              "string",
              "null"
            ]
          },
          "durationMinutes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "isActive": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "maxSessions": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "perMinuteRate": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "planType": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "timeCredits": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "timeWindowEnd": {
            "type": [
              "string",
              "null"
            ]
          },
          "timeWindowStart": {
            "type": [
              "string",
              "null"
            ]
          },
          "validityDays": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "statusCode",
          "message",
          "error",
          "timestamp"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "Plan": {
        "type": "object",
        "required": [
          "id",
          "name",
          "price",
          "planType",
          "durationMinutes",
          "validityDays",
          "timeCredits",
          "perMinuteRate",
          "isActive",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "deviceSubType": {
            "type": [
              "string",
              "null"
            ]
          },
          "deviceType": {
            "type": [
              "string",
              "null"
            ]
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean"
          },
          "maxSessions": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "perMinuteRate": {
            "type": "number",
            "format": "double"
          },
          "planType": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "timeCredits": {
            "type": "integer",
            "format": "int32"
          },
          "timeWindowEnd": {
            "type": [
              "string",
              "null"
            ]
          },
          "timeWindowStart": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "validityDays": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PlanEnvelope": {
        "type": "object",
        "required": [
          "success",
          "statusCode",
          "timestamp",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Plan"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "PlanPaginationEnvelope": {
        "type": "object",
        "required": [
          "success",
          "statusCode",
          "timestamp",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PlanPaginationPage"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
      "PlanPaginationPage": {
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "limit",
          "totalPages"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Plan"
            }
          },
          "limit": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "UpdatePlanDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "deviceSubType": {
            "type": [
              "string",
              "null"
            ]
          },
          "deviceType": {
            "type": [
              "string",
              "null"
            ]
          },
          "durationMinutes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "isActive": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "maxSessions": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "perMinuteRate": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "planType": {
            "type": [
              "string",
              "null"
            ]
          },
          "price": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "timeCredits": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "timeWindowEnd": {
            "type": [
              "string",
              "null"
            ]
          },
          "timeWindowStart": {
            "type": [
              "string",
              "null"
            ]
          },
          "validityDays": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from POST /auth/verify-otp"
      }
    }
  },
...
}